@@ -10,7 +10,7 @@
     ],
     "minimum-stability": "dev",
     "require": {
-        "php": ">=5.4.0"
+        "php": "^7.0 || ^8.0"
     },
     "autoload" : {
 	   "psr-4": {
@@ -18,7 +18,7 @@
        }
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8|^5.3|^6.5",
-        "squizlabs/php_codesniffer": "2.*"
+        "phpunit/phpunit": "^6.5 || ^9.5",
+        "squizlabs/php_codesniffer": "^3.0 || ^3.5"
     }
 }
