@@ -10,7 +10,7 @@ class ContainsTest extends TestCase
      */
     private $col;
 
-    public function setup()
+    public function setup(): void
     {
         $this->col = new Collection('TestClassA',
             [
