@@ -3,7 +3,6 @@
 namespace Collections\Tests\Dictionary;
 
 use Collections\Dictionary;
-use Collections\Exceptions\InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 
 class MergeTest extends TestCase
@@ -43,22 +42,18 @@ class MergeTest extends TestCase
         $this->assertEquals(2, $m->get('a'));
     }
 
-   /**
-     * @expectedException Collections\Exceptions\InvalidArgumentException
-     */
     public function test_merge_of_bad_types_fails()
     {
+        $this->expectException(\Collections\Exceptions\InvalidArgumentException::class);
         $l = (new Dictionary('string','int'))->add('a',1);
         $r = (new Dictionary('string','string'))->add('b','2');
 
         $m = $l->merge($r);
     }
 
-   /**
-     * @expectedException \InvalidArgumentException
-     */
     public function test_merge_of_non_array()
     {
+        $this->expectException(\InvalidArgumentException::class);
         $d = (new Dictionary('string', 'int'))->add('a',1);
 
         $result = $d->merge(3);
